requirements.txt
boto3==1.5.14
botocore==1.8.28
certifi==2018.8.24
chardet==3.0.4
defusedxml==0.5.0
Django==2.1.2
django-allauth==0.38.0
django-appconf==1.0.2
django-cors-headers==2.4.0
django-imagekit==4.0.2
django-render-block==0.5
django-rest-auth==0.9.3
django-rest-framework==0.1.0
django-simple-history==2.4.0
django-storages==1.7.1
django-templated-email==2.2.0
djangorestframework==3.8.2
docutils==0.14
et-xmlfile==1.0.1
idna==2.7
jdcal==1.3
jmespath==0.9.3
mysqlclient==1.3.13
oauthlib==2.1.0
openpyxl==2.4.9
pilkit==2.0
Pillow==5.0.0
python-dateutil==2.6.1
python3-openid==3.1.0
pytz==2018.5
requests==2.19.1
requests-oauthlib==1.0.0
s3transfer==0.1.12
six==1.11.0
urllib3==1.23
